requirements.txt
fastapi
uvicorn
python-multipart
pydantic
openai-whisper
sentence-transformers
ffmpeg-python